training/saas_admin/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7
pydantic-settings==2.1.0
python-dotenv==1.0.0

